corrected n_points

+++ sim_data.py
@@ -42,7 +42,7 @@ print('Config file copied')
 resolution = 4000
 wlen_borders = [2,4]
 stepsize = (wlen_borders[0]+wlen_borders[1])/2/resolution
-N_points = (wlen_borders[1]-wlen_borders[0])/stepsize
+N_points = int((wlen_borders[1]-wlen_borders[0])/stepsize)
 wlen_data = np.linspace(wlen_borders[0],wlen_borders[1],N_points)
 
 instruments = {}
